Add pinWindowTo prop.
By default, when the content size of a scroll view changes across platforms the
top position of the content always stays constant. This makes sense since that’s
usually how scroll position is measured. `element.scrollTop` on web and
`contentOffset.y` on iOS.

However, this behavior is not ideal for chat applications where new content is
being added to the bottom of the scroll view. In a chat app we want to keep the
bottom offset of the scroll window constant instead of the top. The
`pinWindowTo` prop allows this behavior to be configurable!

Demo: https://www.dropbox.com/s/sj3zzjtte72eea9/pinWindowTo-demo.mov?dl=0